requirements.txt
groq
python-dotenv
ffmpeg
pyaudio
speechrecognition
pydub
gtts
elevenlabs
gradio